+++ lizard_portal/media/lizard_portal/lizard/plugin/ApplyContext.js
@@ -5,6 +5,7 @@ Ext.define('Lizard.plugin.ApplyContext', {
         var max_levels_deep = max_levels_deep || 10;
         var parent = object_self.ownerCt;
         for (var i = 0; i < max_levels_deep; i++) {
+            if (typeof parent == 'undefined'){return false;}
             if (parent.events[eventname]) {
                 parent.on(eventname, event_function, object_self);
                 return true;