+++ lizard_portal/testsettings.py
@@ -32,19 +32,23 @@ DATABASES = {
     # the specified database exists. When the tests cannot run, Jenkins sees
     # that as an error.
     'default': {
-        'NAME': os.path.join(BUILDOUT_DIR, 'var', 'sqlite', 'test.db'),
-        'ENGINE': 'django.db.backends.sqlite3',
-        'USER': '',
-        'PASSWORD': '',
-        'HOST': '',  # empty string for localhost.
+        'NAME': 'lizard-portal',
+        'ENGINE': 'django.contrib.gis.db.backends.postgis',
+        'USER': 'buildout',
+        'PASSWORD': 'buildout',
+        'HOST': 'localhost',  # empty string for localhost.
         'PORT': '',  # empty string for default.
         }
     }
+
 SITE_ID = 1
 INSTALLED_APPS = [
+    'lizard_geo',
     'lizard_area',
     'lizard_portal',
     'lizard_security',
+    'djcelery',
+    'lizard_task',
     'lizard_ui',
     'staticfiles',
     'compressor',